+++ pyvisim/neural_networks/siamese/__init__.py
@@ -1,4 +1,4 @@
-from .backbones import ResNetBackbone
+from .backbones import ResNetBackbone, get_transform
 from .bce_siamese_network import BCESiameseNetwork
 from .contrastive_siamese_network import ContrastiveSiameseNetwork
 
@@ -6,4 +6,5 @@ __all__ = [
     "BCESiameseNetwork",
     "ContrastiveSiameseNetwork",
     "ResNetBackbone",
+    "get_transform",
 ]